requirements.txt
requests
pyyaml
paramiko
schedule
python-dotenv